Completes dummy app
