+++ package.json
@@ -13,6 +13,7 @@
   "license": "ISC",
   "dependencies": {
     "@babel/runtime": "^7.15.3",
+    "axios": "^0.21.1",
     "express": "^4.17.1",
     "mongoose": "^5.13.7",
     "nodemon": "^2.0.12",
